create CompatAsyncTask.kt, use in test, some logic fixes
