+++ sample-app/src/androidTest/java/com/atiurin/sampleapp/tests/espresso/ViewTest.kt
@@ -5,7 +5,8 @@ import androidx.test.core.app.ActivityScenario
 import androidx.test.espresso.matcher.ViewMatchers.withId
 import com.atiurin.sampleapp.R
 import com.atiurin.sampleapp.activity.CustomClicksActivity
-import com.atiurin.sampleapp.activity.CustomClicksActivity.Companion.COMPAT_ASYNC_TASK_TIME_EXECUTION
+import com.atiurin.sampleapp.async.task.CompatAsyncTask
+import com.atiurin.sampleapp.async.task.CompatAsyncTask.Companion.COMPAT_ASYNC_TASK_TIME_EXECUTION
 import com.atiurin.sampleapp.tests.BaseTest
 import com.atiurin.ultron.custom.espresso.action.getView
 import com.atiurin.ultron.custom.espresso.base.getViewForcibly
@@ -13,21 +14,24 @@ import com.atiurin.ultron.extensions.isChecked
 import com.atiurin.ultron.extensions.performOnView
 import com.atiurin.ultron.extensions.performOnViewForcibly
 import com.atiurin.ultron.testlifecycle.setupteardown.SetUpRule
+import com.atiurin.ultron.testlifecycle.setupteardown.TearDown
+import com.atiurin.ultron.testlifecycle.setupteardown.TearDownRule
 import org.junit.Assert
 import org.junit.Test
 
 class ViewTest : BaseTest() {
 
-    private lateinit var customClicksActivity: CustomClicksActivity
+    private val compatAsyncTask = CompatAsyncTask()
 
-    private val startActivity = SetUpRule().add {
-        ActivityScenario.launch(CustomClicksActivity::class.java).onActivity { activity ->
-            customClicksActivity = activity
-        }
+    private val startActivity = SetUpRule()
+        .add { ActivityScenario.launch(CustomClicksActivity::class.java) }
+
+    private val tearDownRule = TearDownRule().add(STOP_ASYNC_TASK) {
+        compatAsyncTask.stop()
     }
 
     init {
-        ruleSequence.addLast(startActivity)
+        ruleSequence.add(tearDownRule).addLast(startActivity)
     }
 
     @Test
@@ -37,9 +41,10 @@ class ViewTest : BaseTest() {
     }
 
     @Test
+    @TearDown(STOP_ASYNC_TASK)
     fun actionGetViewForcibly() {
         val startTime = System.currentTimeMillis()
-        customClicksActivity.startCompatAsyncTask()
+        compatAsyncTask.start()
         val view = withId(R.id.rB_top_right).getViewForcibly()
         Assert.assertNotNull(view)
         Assert.assertTrue(System.currentTimeMillis() < COMPAT_ASYNC_TASK_TIME_EXECUTION + startTime)
@@ -57,9 +62,10 @@ class ViewTest : BaseTest() {
     }
 
     @Test
+    @TearDown(STOP_ASYNC_TASK)
     fun matcherActionPerformOnViewForcibly() {
         val startTime = System.currentTimeMillis()
-        customClicksActivity.startCompatAsyncTask()
+        compatAsyncTask.start()
         withId(R.id.rB_top_left).apply {
             performOnViewForcibly {
                 performClick()
@@ -67,5 +73,9 @@ class ViewTest : BaseTest() {
             }
         }
         Assert.assertTrue(System.currentTimeMillis() < COMPAT_ASYNC_TASK_TIME_EXECUTION + startTime)
+    }
+
+    companion object {
+        const val STOP_ASYNC_TASK = "STOP_ASYNC_TASK"
     }
 }
