+++ ultron/src/main/java/com/atiurin/ultron/extensions/PerfomOnViewExt.kt
@@ -8,7 +8,7 @@ import com.atiurin.ultron.core.espresso.UltronEspressoInteraction
 import com.atiurin.ultron.custom.espresso.action.CustomEspressoActionType.PERFORM_ON_VIEW
 import com.atiurin.ultron.custom.espresso.action.CustomEspressoActionType.PERFORM_ON_VIEW_FORCIBLY
 import com.atiurin.ultron.custom.espresso.action.getView
-import com.atiurin.ultron.custom.espresso.base.getViewForcibly
+import com.atiurin.ultron.custom.espresso.base.UltronViewFinder
 import com.atiurin.ultron.utils.runOnUiThread
 import org.hamcrest.Matcher
 
@@ -23,8 +23,9 @@ fun View.performOnView(action: View.() -> Unit) {
 }
 
 fun <T> UltronEspressoInteraction<T>.performOnView(action: View.() -> Unit) {
+    val view = this.getView()
     executeAction(
-        operationBlock = { this.getView().performOnView(action) },
+        operationBlock = { view.performOnView(action) },
         name = "Perform on view with '${getInteractionMatcher()}' in root '${getInteractionRootMatcher()}'",
         type = PERFORM_ON_VIEW,
         description = "${interaction.className()} perform $action on view '$PERFORM_ON_VIEW' of '${getInteractionMatcher()}' with root '${getInteractionRootMatcher()}' during ${getActionTimeout()} ms",
@@ -33,7 +34,7 @@ fun <T> UltronEspressoInteraction<T>.performOnView(action: View.() -> Unit) {
 
 fun <T> UltronEspressoInteraction<T>.performOnViewForcibly(action: View.() -> Unit) {
     executeAction(
-        operationBlock = { this.getViewForcibly().performOnView(action) },
+        operationBlock = { UltronViewFinder(interaction).view.performOnView(action) },
         name = "Perform forcibly on view with '${getInteractionMatcher()}' in root '${getInteractionRootMatcher()}'",
         type = PERFORM_ON_VIEW_FORCIBLY,
         description = "${interaction.className()} perform $action forcibly on view '$PERFORM_ON_VIEW_FORCIBLY' of '${getInteractionMatcher()}' with root '${getInteractionRootMatcher()}' during ${getActionTimeout()} ms",
